# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample service for managing orders",
        "title": "Orders API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "soberkoder@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8082",
    "paths": {
        "/subscription/create": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Подписка"
                ],
                "summary": "Создание подписки",
                "parameters": [
                    {
                        "description": "Подписка",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateSubscriptionDto"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Ошибка валидации",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorValidate"
                        }
                    },
                    "500": {
                        "description": "Другие ошибки",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "constant.ResponseMessage": {
            "type": "string",
            "enum": [
                "VALIDATION_ERROR"
            ],
            "x-enum-varnames": [
                "ValidationError"
            ]
        },
        "helper.ErrorResponse": {
            "type": "object",
            "properties": {
                "errorMessage": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "helper.ErrorValidate": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "additionalProperties": {
                            "$ref": "#/definitions/helper.Field"
                        }
                    }
                },
                "message": {
                    "$ref": "#/definitions/constant.ResponseMessage"
                }
            }
        },
        "helper.Field": {
            "type": "object",
            "properties": {
                "messages": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "model.CreateSubscriptionDto": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        }
    }
}